% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amazonas_HQoL.R
\docType{data}
\name{Amazonas_HQoL}
\alias{Amazonas_HQoL}
\title{Amazonas_HQoL: Data on Health related Quality of Life and Malaria in Brazilian Amazonas}
\format{
A dataframe with 1179 rows and 47 columns:
\describe{
\item{\code{P77}}{One of the outcomes, Quality of life measure, Visual Analog Scale (VAS), continous }
\item{\code{FILTRO}}{Main exposure, categorical, whether the individual is part of a “treatment” (recent malaria) or control (past malaria)}
}
}
\source{
\url{https://doi.org/10.1371/journal.pntd.0012739}
}
\usage{
Amazonas_HQoL
}
\description{
This data is from a publicly available 'open data' set from a peer reviewed manuscript. The study estimated health-related quality of life (HRQoL) losses due to malaria in endemic areas of the Brazilian Amazon using the EQ-5D-3L instrument. Data is from a convenience sample of 1,179 individuals aged 18 years or older.

Reference Source:  Andrade MV, de Souza Noronha KVM, Guedes GR, Julião NA, de Carvalho LR, de Souza A, et al. (2024) Health-Related Quality of Life due to malaria in the Brazilian Amazon using EQ-5D-3L. PLoS Negl Trop Dis 18(12): e0012739. https://doi.org/10.1371/journal.pntd.0012739
}
\keyword{datasets}
