% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/births.R
\docType{data}
\name{births}
\alias{births}
\title{Births: Data of ...}
\format{
A dataframe with 1000 rows and 6 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{... pending url}
}
\usage{
births
}
\description{
extra details pending ...
}
\keyword{datasets}
