% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Covid_Bangladesh.R
\docType{data}
\name{Covid_Bangladesh}
\alias{Covid_Bangladesh}
\title{Covid_Bangladesh: Data of Covid-19 vaccines coverage in Bangladesh}
\format{
A dataframe with 2502 rows and 24 columns:
\describe{
\item{vdoses}{Main outcome, Number of COVID-19 doses, count, discrete}
\item{wbscore}{Mental wellbeing score (WHO), continuous}
\item{chroill_1}{Presence of chronic illness,	binary}
\item{mi_1}{Whether the individual migrated within last 12 months, binary}
}
}
\source{
\url{https://doi.org/10.1371/journal.pgph.0004178}
}
\usage{
Covid_Bangladesh
}
\description{
This data is from a publicly available 'open data' set from a peer reviewed manuscript. The study was a cross-sectional study conducted between August and September 2022 in Duaripara, a slum in northeast Dhaka (in-migration site), and Tala, a disaster-prone sub-district in southwest Satkhira (out-migration site). Data included information from 1,239 adults in Duaripara and 1,263 adults in Tala from 625 and 596 randomly selected households, respectively.

Reference Source:  Islam MN, De Allegri M, Bonnet E, Sarker M, Goudet J-M, Franceschin L, et al. (2025) High coverage and equitable distribution of COVID-19 vaccine uptake in two vulnerable areas in Bangladesh. PLOS Glob Public Health 5(1): e0004178. https://doi.org/10.1371/journal.pgph.0004178
}
\keyword{datasets}
