% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covidkenya.R
\docType{data}
\name{covidkenya}
\alias{covidkenya}
\title{covidkenya: Data of ...}
\format{
A dataframe with 355 rows and 15 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{https://doi.org/10.1371/journal.pgph.0002011}
}
\usage{
covidkenya
}
\description{
Source:  Tsegaye AT, Sherry C, Oduol C, Otieno J, Rwigi D, et al. (2023) Clinical epidemiology of COVID-19 among hospitalized children in rural western Kenya. PLOS Global Public Health 3(6): e0002011. https://doi.org/10.1371/journal.pgph.0002011
}
\keyword{datasets}
