% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_pd_sep22pm.R
\docType{data}
\name{db_pd_sep22pm}
\alias{db_pd_sep22pm}
\title{db_pd_sep22pm: Data of ...}
\format{
A dataframe with 839 rows and 520 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{... pending url}
}
\usage{
db_pd_sep22pm
}
\description{
extra details pending ...
}
\keyword{datasets}
