% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epib.704.data-package.R
\docType{package}
\name{epib.704.data-package}
\alias{epib.704.data}
\alias{epib.704.data-package}
\title{epib.704.data: Data for EPIB 704 Assignments}
\description{
This R package contains datasets specifically for use in EPIB 704 assignments, including data related to various epidemiological studies and assignments.
}
\author{
\strong{Maintainer}: Edgar Ortiz-Brizuela \email{edgar.ortiz-brizuela@mail.mcgill.ca} (\href{https://orcid.org/0000-0001-7169-8459}{ORCID})

}
\keyword{internal}
