% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plato.R
\docType{data}
\name{plato}
\alias{plato}
\title{Plato: Data of ...}
\format{
A dataframe with 6 rows and 5 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{... pending url}
}
\usage{
plato
}
\description{
extra details pending ...
}
\keyword{datasets}
