% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampledata.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{sampledata: Data of ...}
\format{
A dataframe with 272 rows and 3 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{... pending url}
}
\usage{
sampledata
}
\description{
extra details pending ...
}
\keyword{datasets}
