% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sbdata1.R
\docType{data}
\name{sbdata1}
\alias{sbdata1}
\title{sbdata1: Data of ...}
\format{
A dataframe with 839 rows and 29 columns:
\describe{
\item{columna1}{description}
\item{columna2}{description}
}
}
\source{
\url{... pending url}
}
\usage{
sbdata1
}
\description{
extra details pending ...
}
\keyword{datasets}
